Moved Custom_Thread class to separate file
